+++ example-app/app/clients/index.ts
@@ -1,0 +1,3 @@
+export { PublicClient } from './PublicClient'
+export { WalletClient } from './WalletClient'
+export { WebsocketClient } from './WebsocketClient'
