+++ example-app/tsconfig.json
@@ -5,7 +5,8 @@
     "jsx": "react-jsx",
     "paths": {
       "@/transfer": ["./app/transfer"],
-      "@/components": ["./app/components"]
+      "@/components": ["./app/components"],
+      "@/clients": ["./app/clients"]
     }
   }
 }
